requirements.txt
aiogram==2.25
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
black==24.1.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
mypy-extensions==1.0.0
openai==1.10.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pydantic==2.6.0
pydantic_core==2.16.1
python-dotenv==1.0.1
pytz==2023.4
sniffio==1.3.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.9.0
yarl==1.9.4
